# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waypoint Delivery Planning System API
  description: OpenAPI 3.1 specification for the Waypoint Group delivery planning system.
  contact:
    email: tech-triathlon@rootcode.io
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production API Base Endpoint
...
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: User login to receive JWT bearer token
      tags:
        - auth
      x-role:
        - dispatcher
        - loader
        - driver
        - store_manager
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: dispatcher@waypoint.test
              password: Password123!
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example
                token_type: bearer
                user:
                  id: USR001
                  username: dispatcher@waypoint.test
                  role: dispatcher
                  depot_ids:
                    - Peliyagoda
                    - Kandy
                  display_name: Lead Dispatcher
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Invalid username or password
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      operationId: getCurrentUser
      summary: Get current authenticated user profile and scopes
      tags:
        - auth
      x-role:
        - dispatcher
        - loader
        - driver
        - store_manager
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: USR001
                username: dispatcher@waypoint.test
                role: dispatcher
                depot_ids:
                  - Peliyagoda
                  - Kandy
                display_name: Lead Dispatcher
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer authentication header. Role and scope are derived from the token.
  schemas:
...
    Role:
      type: string
      enum:
        - dispatcher
        - loader
        - driver
        - store_manager
      description: User role within the Waypoint delivery planning system.
...
    Violation:
      type: object
      required:
        - rule
        - message
      properties:
        rule:
          type: string
          description: Rule ID (e.g. H1 through H12).
        message:
          type: string
          description: Human-readable description of rule violation.
        actual:
          description: Observed value that violated constraint.
        limit:
          description: Threshold limit specified by the rule.
      description: Constraint violation detail item.
    ErrorDetails:
      type: object
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          description: List of rule violations if code is CONSTRAINT_VIOLATION.
      additionalProperties: true
      description: Detailed contextual metadata for error payload.
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g. CONSTRAINT_VIOLATION, PLAN_CHANGED, NOT_OPERATING_DAY, INVALID_TRANSITION, FORBIDDEN_SCOPE, OPEN_SHORTFALL, STALE_PLAN_VERSION).
        message:
          type: string
          description: High-level error message.
        details:
          $ref: '#/components/schemas/ErrorDetails'
      description: Core error payload.
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      description: Standard API error response envelope.

    # ----------------------------------------------------
    # Domain Model & DTO Schemas
    # ----------------------------------------------------
...
    User:
      type: object
      required:
        - id
        - username
        - role
        - display_name
      properties:
        id:
          type: string
          description: User ID.
        username:
          type: string
          description: Login username / email address.
        role:
          $ref: '#/components/schemas/Role'
        depot_ids:
          type: array
          items:
            type: string
          description: Depot scoping list for dispatcher or loader (D24).
        outlet_id:
          type: string
          nullable: true
          description: Outlet scoping for store manager.
        vehicle_id:
          type: string
          nullable: true
          description: Vehicle scoping for driver.
        display_name:
          type: string
          description: Human-readable display name.
      description: System user account model.
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
      description: User credentials for login endpoint.
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        user:
          $ref: '#/components/schemas/User'
      description: Authentication token payload response.
...